# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/contracts": {
      "get": {
        "operationId": "ContractsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "合同包与模拟签署"
        ]
      },
      "post": {
        "operationId": "ContractsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContract"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "合同包与模拟签署"
        ]
      }
    },
    "/api/contracts/confirmations": {
      "get": {
        "operationId": "ContractsController_confirmations",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "合同包与模拟签署"
        ]
      }
    },
    "/api/contracts/templates": {
      "get": {
        "operationId": "ContractsController_templates",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "合同包与模拟签署"
        ]
      },
      "post": {
        "operationId": "ContractsController_template",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "合同包与模拟签署"
        ]
      }
    },
    "/api/contracts/templates/{id}/publish": {
      "post": {
        "operationId": "ContractsController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "合同包与模拟签署"
        ]
      }
    },
    "/api/contracts/{id}": {
      "get": {
        "operationId": "ContractsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "合同包与模拟签署"
        ]
      },
      "put": {
        "operationId": "ContractsController_edit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeContract"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "合同包与模拟签署"
        ]
      }
    },
    "/api/contracts/{id}/archive": {
      "get": {
        "operationId": "ContractsController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "合同包与模拟签署"
        ]
      }
    },
    "/api/contracts/{id}/change": {
      "post": {
        "operationId": "ContractsController_change",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeContract"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "合同包与模拟签署"
        ]
      }
    },
    "/api/contracts/{id}/{action}": {
      "post": {
        "operationId": "ContractsController_action",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionAction"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "合同包与模拟签署"
        ]
      }
    },
    "/api/fulfillment": {
      "get": {
        "operationId": "FulfillmentController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "联运执行"
        ]
      },
      "post": {
        "operationId": "FulfillmentController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "联运执行"
        ]
      }
    },
    "/api/fulfillment/resources": {
      "get": {
        "operationId": "FulfillmentController_resources",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "联运执行"
        ]
      }
    },
    "/api/fulfillment/vehicles": {
      "post": {
        "operationId": "FulfillmentController_vehicle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "联运执行"
        ]
      }
    },
    "/api/fulfillment/evidence/{id}": {
      "get": {
        "operationId": "FulfillmentController_download",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "联运执行"
        ]
      }
    },
    "/api/fulfillment/{id}": {
      "get": {
        "operationId": "FulfillmentController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "联运执行"
        ]
      }
    },
    "/api/fulfillment/{id}/tasks": {
      "post": {
        "operationId": "FulfillmentController_dispatch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "联运执行"
        ]
      }
    },
    "/api/fulfillment/tasks/{id}/events": {
      "post": {
        "operationId": "FulfillmentController_event",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "联运执行"
        ]
      }
    },
    "/api/fulfillment/tasks/{id}/cancel": {
      "post": {
        "operationId": "FulfillmentController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "联运执行"
        ]
      }
    },
    "/api/fulfillment/tasks/{id}/issues": {
      "post": {
        "operationId": "FulfillmentController_issue",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "联运执行"
        ]
      }
    },
    "/api/fulfillment/tasks/{id}/evidence": {
      "post": {
        "operationId": "FulfillmentController_upload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "联运执行"
        ]
      }
    },
    "/api/fulfillment/issues/{id}/resolve": {
      "post": {
        "operationId": "FulfillmentController_resolve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolutionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "联运执行"
        ]
      }
    },
    "/api/fulfillment/{id}/batches": {
      "post": {
        "operationId": "FulfillmentController_batch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "联运执行"
        ]
      }
    },
    "/api/fulfillment/batches/{id}/arrivals": {
      "post": {
        "operationId": "FulfillmentController_arrival",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArrivalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "联运执行"
        ]
      }
    },
    "/api/fulfillment/batches/{id}/boxes": {
      "post": {
        "operationId": "FulfillmentController_box",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoxDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "联运执行"
        ]
      }
    },
    "/api/driver/login": {
      "post": {
        "operationId": "DriverController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "司机电脑模拟端"
        ]
      }
    },
...
    "/api/driver/password": {
      "post": {
        "operationId": "DriverController_password",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "司机电脑模拟端"
        ]
      }
    },
...
    "/api/driver/tasks/{id}/events": {
      "post": {
        "operationId": "DriverController_event",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "司机电脑模拟端"
        ]
      }
    },
    "/api/driver/tasks/{id}/issues": {
      "post": {
        "operationId": "DriverController_issue",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "司机电脑模拟端"
        ]
      }
    },
    "/api/driver/tasks/{id}/evidence": {
      "post": {
        "operationId": "DriverController_upload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "司机电脑模拟端"
        ]
      }
    },
    "/api/driver/evidence/{id}": {
      "get": {
        "operationId": "DriverController_download",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "司机电脑模拟端"
        ]
      }
    },
...
  },
  "info": {
    "title": "北粮南运 · 智能多式联运 API",
    "description": "Phase 9：需求、方案、匹配、模拟签章、联运执行、预测预警、智能辅助、账单版本、双方结算确认和授权数据服务。",
    "version": "0.9.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TemplateDto": {
        "type": "object",
        "properties": {}
      },
      "CreateContract": {
        "type": "object",
        "properties": {}
      },
      "ChangeContract": {
        "type": "object",
        "properties": {}
      },
      "RevisionAction": {
        "type": "object",
        "properties": {}
      },
      "BusinessDto": {
        "type": "object",
        "properties": {}
      },
      "VehicleDto": {
        "type": "object",
        "properties": {}
      },
      "TaskDto": {
        "type": "object",
        "properties": {}
      },
      "EventDto": {
        "type": "object",
        "properties": {}
      },
      "IssueDto": {
        "type": "object",
        "properties": {}
      },
      "ResolutionDto": {
        "type": "object",
        "properties": {}
      },
      "BatchDto": {
        "type": "object",
        "properties": {}
      },
      "ArrivalDto": {
        "type": "object",
        "properties": {}
      },
      "BoxDto": {
        "type": "object",
        "properties": {}
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "PasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
...
    }
  }
}